prova: Nonparametric Probabilistic-Statistical Variate Analysis with Automated Markov-Chain Monte Carlo

Functions for calculating posterior joint and conditional probabilities, probability distributions of population frequencies, and information-theoretic measures, by means of Bayesian nonparametric methods. Markov-chain Monte Carlo calculations are automatically handled and do not require user supervision. Applications range from statistical estimation and probabilistic hypothesis testing to evidence-based inference and decision making, in a wide range of disciplines from astrophysics to medicine. For more details and examples see for instance Porta Mana et al. (2026) doi:10.31219/osf.io/8nr56 , Dunson & Bhattacharya (2011) doi:10.1093/acprof:oso/9780199694587.003.0005 , Lindley & Novick (1981) doi:10.1214/aos/1176345331 , Bernardo & Smith (2000) doi:10.1002/9780470316870 , Müller et al. (2015) doi:10.1007/978-3-319-18968-0 . Requires the packages 'Nimble', 'parallel', 'extraDistr'.

See also

Author

Maintainer: PierGianLuca Porta Mana pgl@portamana.org (ORCID) [copyright holder]

Authors:

PierGianLuca Porta Mana pgl@portamana.org (ORCID) [copyright holder]

Other contributors:

Aurora Grefsrud agre@hvl.no (ORCID) [contributor]

Håkon Mydland haakon.mydland@gmail.com (ORCID) [contributor]

Maksim Ohvrill maksimohvrill@hotmail.com [contributor]

Simen Hesthamar Hauge simen@hhnet.no (ORCID) [contributor]